--- kadai3-2/1tsuki/README.md
@@ -1,6 +1,0 @@
-# 分割ダウンロードを行う
-- [x] Rangeアクセスを用いる
-- [x] いくつかのゴルーチンでダウンロードしてマージする
-- [ ] エラー処理を工夫する
-- [ ] golang.org/x/sync/errgourpパッケージなどを使ってみる
-- [x] キャンセルが発生した場合の実装を行う
