+++ kadai3-2/1tsuki/pget/README.md
@@ -1,0 +1,6 @@
+# 分割ダウンロードを行う
+- [x] Rangeアクセスを用いる
+- [x] いくつかのゴルーチンでダウンロードしてマージする
+- [ ] エラー処理を工夫する
+- [ ] golang.org/x/sync/errgourpパッケージなどを使ってみる
+- [x] キャンセルが発生した場合の実装を行う
